refactor: split Android home lifecycle actions
